+++ product-service/src/main/java/com/mall/productservice/repository/CartItemRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CartItemRepository extends JpaRepository<CartItem, Long> {
-    List<CartItem> findByUserId(Long userId);
-} 
+    List<CartItem> findByUserId(Integer userId);
+}
